Moved the (bad) experimental screening function into the file with the algorithm and just added an enum for basic/screened mode. Apart from this step, the two algorithms are exactly the same
